refactor: helpers for countsmap tests
