+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/utils/CountsMapTest.kt
@@ -1,10 +1,9 @@
 package xyz.lbres.kotlinutils.set.multiset.utils
 
+import xyz.lbres.kotlinutils.set.multiset.utils.countsmap.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.Test
 import kotlin.test.assertContains
 import kotlin.test.assertEquals
-import kotlin.test.assertFalse
-import kotlin.test.assertTrue
 
 class CountsMapTest {
     @Test
@@ -17,149 +16,11 @@ class CountsMapTest {
         // TODO
     }
 
-    @Test
-    fun testGetCountOf() {
-        var intMap: CountsMap<Int> = CountsMap.from(emptyList())
-        assertEquals(0, intMap.getCountOf(0))
-        assertEquals(0, intMap.getCountOf(100))
+    @Test fun testGetCountOf() = runGetCountOfTests()
+    @Test fun testIsEmpty() = runIsEmptyTests()
 
-        intMap = CountsMap.from(listOf(2))
-        assertEquals(1, intMap.getCountOf(2))
-        assertEquals(0, intMap.getCountOf(1))
-
-        intMap = CountsMap.from(listOf(1, 1, 2, 1, -4, 5, 2))
-        assertEquals(3, intMap.getCountOf(1))
-        assertEquals(2, intMap.getCountOf(2))
-        assertEquals(1, intMap.getCountOf(-4))
-        assertEquals(1, intMap.getCountOf(5))
-
-        val listMap = CountsMap.from(listOf(listOf(1, 2, 3), listOf(1, 2, 3)))
-        assertEquals(2, listMap.getCountOf(listOf(1, 2, 3)))
-        assertEquals(0, listMap.getCountOf(listOf(1, 2)))
-    }
-
-    @Test
-    fun testIsEmpty() {
-        // empty
-        var intMap: CountsMap<Int> = CountsMap.from(emptyList())
-        assertTrue(intMap.isEmpty())
-
-        var stringMap: CountsMap<String> = CountsMap.from(emptyList())
-        assertTrue(stringMap.isEmpty())
-
-        // not empty
-        intMap = CountsMap.from(listOf(0))
-        assertFalse(intMap.isEmpty())
-
-        intMap = CountsMap.from(listOf(1000, -1000, 4, 2, 4))
-        assertFalse(intMap.isEmpty())
-
-        intMap = CountsMap.from(listOf(3, 3, 3))
-        assertFalse(intMap.isEmpty())
-
-        stringMap = CountsMap.from(listOf("123", "abc"))
-        assertFalse(stringMap.isEmpty())
-
-        stringMap = CountsMap.from(listOf("hello world", "hello world"))
-        assertFalse(stringMap.isEmpty())
-    }
-
-    @Test
-    fun testContains() {
-        var intMap: CountsMap<Int> = CountsMap.from(emptyList())
-        assertFalse(intMap.contains(0))
-        assertFalse(intMap.contains(1000))
-        assertFalse(intMap.contains(-1000))
-
-        intMap = CountsMap.from(listOf(1, 2))
-        assertFalse(intMap.contains(0))
-        assertTrue(intMap.contains(1))
-        assertTrue(intMap.contains(2))
-
-        intMap = CountsMap.from(listOf(1, 1, 1))
-        assertTrue(intMap.contains(1))
-        assertFalse(intMap.contains(2))
-
-        val error = ArithmeticException()
-        val errorMap = CountsMap.from(listOf(ArithmeticException(), error, NumberFormatException()))
-        assertTrue(errorMap.contains(error))
-
-        val listMap = CountsMap.from(listOf(emptyList(), listOf(5, 6), listOf(9, 8, 3)))
-        assertTrue(listMap.contains(emptyList()))
-        assertTrue(listMap.contains(listOf(9, 8, 3)))
-        assertFalse(listMap.contains(listOf(6, 6)))
-    }
-
-    @Test
-    fun testContainsAll() {
-        // equal
-        var map: CountsMap<Int> = CountsMap.from(emptyList())
-        var collection: Collection<Int> = emptyList()
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(-445))
-        collection = listOf(-445)
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(1, 1))
-        collection = listOf(1, 1)
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(2, 3, 2, 4, 3, 4, 4))
-        collection = listOf(2, 3, 2, 4, 3, 4, 4)
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(1, 2, 3))
-        collection = listOf(3, 1, 2)
-        assertTrue(map.containsAll(collection))
-
-        // submap
-        map = CountsMap.from(listOf(1))
-        collection = emptyList()
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(emptyList())
-        collection = listOf(1)
-        assertFalse(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(1, 2, 3, 4))
-        collection = listOf(1, 3)
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(1, 1, 1))
-        collection = listOf(1, 1)
-        assertTrue(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(1, 3, 5))
-        collection = listOf(1, 3, -1, 5)
-        assertFalse(map.containsAll(collection))
-
-        // overlapping keys
-        map = CountsMap.from(listOf(1, 2, 3))
-        collection = listOf(1, 3, 4)
-        assertFalse(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(100, 100, 300, 400))
-        collection = listOf(100, 300, 400, 400)
-        assertFalse(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(-10, 5, -10, -10))
-        collection = listOf(-10, -5, -10, -10)
-        assertFalse(map.containsAll(collection))
-
-        // no overlapping keys
-        map = CountsMap.from(listOf(1))
-        collection = listOf(2)
-        assertFalse(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(1, 1, 1, 1))
-        collection = listOf(2, 2, 2, 2)
-        assertFalse(map.containsAll(collection))
-
-        map = CountsMap.from(listOf(4, -4, 5, 7))
-        collection = listOf(22, 23, 22)
-        assertFalse(map.containsAll(collection))
-    }
+    @Test fun testContains() = runContainsTests()
+    @Test fun testContainsAll() = runContainsAllTests()
 
     @Test
     fun testForEach() {
